Commit nueva interfaz
Botones y fuente/color de letra cambiados
Pantallas de registro eliminadas y pantalla de login como principal
