--- app/src/main/java/com/pmydm/projecterato/RegisterActivity.kt
@@ -1,46 +1,0 @@
-package com.pmydm.projecterato
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import android.widget.ImageButton
-import androidx.appcompat.app.AppCompatActivity
-
-class RegisterActivity : AppCompatActivity() {
-    private var volumen=true
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContentView(R.layout.activity_register)
-
-        val imageButtonVolumen: ImageButton = findViewById(R.id.imageButtonVolumen)
-
-        imageButtonVolumen.setOnClickListener {
-            if (volumen) {
-                imageButtonVolumen.setImageResource(R.drawable.iconovolumenapagado)
-                volumen=false
-            } else {
-                imageButtonVolumen.setImageResource(R.drawable.iconovolumenencendido)
-                volumen=true
-            }
-            // Alternar el estado
-        }
-
-        val imageButtonVolver: ImageButton = findViewById(R.id.imageButtonVolver)
-
-        imageButtonVolver.setOnClickListener {
-            val intent = Intent(this, ProfileMenuNLActivity::class.java)
-            startActivity(intent)
-        }
-
-
-
-        val buttonRegister: Button = findViewById(R.id.buttonRegister)
-
-        buttonRegister.setOnClickListener {
-            MainActivity.setLogin()
-            val intent = Intent(this, ProfileMenuActivity::class.java)
-            startActivity(intent)
-        }
-    }
-}
