--- app/src/main/res/drawable/button_background.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#FFB4A2" />
-    <corners android:radius="20dp" />
-</shape>
