+++ app/src/main/res/layout/activity_game.xml
@@ -15,7 +15,8 @@
         android:layout_height="63dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolumen"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -27,8 +28,9 @@
         android:layout_height="63dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolver"
+        android:paddingBottom="8dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@drawable/flecha"
@@ -38,8 +40,12 @@
         android:id="@+id/buttonWin"
         android:layout_width="280dp"
         android:layout_height="252dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:paddingBottom="35dp"
         android:text="@string/click_me_to_win"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:textColor="#6F5D47"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
