+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,11 @@
         android:layout_marginStart="110dp"
         android:layout_marginEnd="110dp"
         android:layout_marginBottom="84dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/banderas"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
@@ -29,7 +33,11 @@
         android:layout_marginStart="110dp"
         android:layout_marginEnd="110dp"
         android:layout_marginBottom="24dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/capitales"
         app:layout_constraintBottom_toTopOf="@+id/buttonBanderas"
         app:layout_constraintEnd_toEndOf="parent"
@@ -42,7 +50,11 @@
         android:layout_marginStart="110dp"
         android:layout_marginEnd="110dp"
         android:layout_marginBottom="24dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/paises"
         app:layout_constraintBottom_toTopOf="@+id/buttonCapitales"
         app:layout_constraintEnd_toEndOf="parent"
@@ -54,7 +66,8 @@
         android:layout_height="63dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolumen"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -66,7 +79,8 @@
         android:layout_height="63dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonmenu"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
