+++ app/src/main/res/layout/activity_menu_region_selection.xml
@@ -16,7 +16,11 @@
         android:layout_marginStart="96dp"
         android:layout_marginTop="116dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/europa"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -30,7 +34,11 @@
         android:layout_marginStart="96dp"
         android:layout_marginTop="40dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/asia"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -44,7 +52,11 @@
         android:layout_marginStart="96dp"
         android:layout_marginTop="40dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/america"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -58,7 +70,11 @@
         android:layout_marginStart="96dp"
         android:layout_marginTop="40dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/africa"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -72,7 +88,11 @@
         android:layout_marginStart="96dp"
         android:layout_marginTop="40dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         android:text="@string/oceania"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -85,7 +105,8 @@
         android:layout_height="63dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolumen"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -97,7 +118,8 @@
         android:layout_height="63dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolver"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
