+++ app/src/main/res/layout/activity_profile_menu.xml
@@ -10,55 +10,88 @@
     android:scaleType="centerCrop">
 
     <Button
-        android:id="@+id/buttonIdioma"
+        android:id="@+id/buttonCerrarSesion"
         android:layout_width="217dp"
         android:layout_height="87dp"
-        android:layout_marginStart="96dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginStart="97dp"
+        android:layout_marginTop="112dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
-        android:text="@string/idioma"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:paddingBottom="9dp"
+        android:text="@string/cerrar_sesion"
+        android:textColor="#6F5D47"
+        android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.47"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/buttonAjustes" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/buttonFondo"
         android:layout_width="217dp"
         android:layout_height="87dp"
         android:layout_marginStart="96dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="28dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:paddingBottom="9dp"
         android:text="@string/cambiar_fondo"
+        android:textColor="#6F5D47"
+        android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.47"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/buttonDescargarDatos" />
 
     <Button
+        android:id="@+id/buttonIdioma"
+        android:layout_width="217dp"
+        android:layout_height="87dp"
+        android:layout_marginStart="96dp"
+        android:layout_marginTop="28dp"
+        android:layout_marginEnd="97dp"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:paddingBottom="9dp"
+        android:text="@string/idioma"
+        android:textColor="#6F5D47"
+        android:textSize="18sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.47"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/buttonAjustes" />
+
+    <Button
         android:id="@+id/buttonAjustes"
         android:layout_width="217dp"
         android:layout_height="87dp"
         android:layout_marginStart="97dp"
-        android:layout_marginTop="304dp"
+        android:layout_marginTop="36dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:paddingBottom="9dp"
         android:text="@string/ajustes"
+        android:textColor="#6F5D47"
+        android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/buttonCerrarSesion" />
 
     <Button
         android:id="@+id/buttonDescargarDatos"
         android:layout_width="217dp"
         android:layout_height="87dp"
         android:layout_marginStart="96dp"
-        android:layout_marginTop="32dp"
+        android:layout_marginTop="36dp"
         android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
+        android:paddingBottom="9dp"
         android:text="@string/descargar_datos"
+        android:textColor="#6F5D47"
+        android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.47"
         app:layout_constraintStart_toStartOf="parent"
@@ -70,7 +103,8 @@
         android:layout_height="63dp"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolver"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -83,7 +117,8 @@
         android:layout_height="63dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolumen"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
