--- app/src/main/res/layout/activity_profile_menu_nl.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ProfileMenuNLActivity"
-    android:background="@drawable/fondoapp"
-    android:scaleType="centerCrop">
-
-    <Button
-        android:id="@+id/buttonLogin"
-        android:layout_width="218dp"
-        android:layout_height="87dp"
-        android:layout_marginStart="96dp"
-        android:layout_marginTop="116dp"
-        android:layout_marginEnd="97dp"
-        android:text="@string/iniciar_sesion"
-        android:backgroundTint="#FFB4A2"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/buttonRegister"
-        android:layout_width="218dp"
-        android:layout_height="87dp"
-        android:layout_marginStart="96dp"
-        android:layout_marginTop="240dp"
-        android:layout_marginEnd="97dp"
-        android:backgroundTint="#FFB4A2"
-        android:text="@string/registrarse"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/imageButtonVolver"
-        android:layout_width="63dp"
-        android:layout_height="63dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
-        android:contentDescription="@string/botonvolver"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/flecha"
-        tools:ignore="DuplicateSpeakableTextCheck" />
-
-    <ImageButton
-        android:id="@+id/imageButtonVolumen"
-        android:layout_width="63dp"
-        android:layout_height="63dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
-        android:contentDescription="@string/botonvolumen"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/iconovolumenencendido" />
-</androidx.constraintlayout.widget.ConstraintLayout>
