--- app/src/main/res/layout/activity_register.xml
@@ -1,133 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".RegisterActivity"
-    android:background="@drawable/fondoapp"
-    android:scaleType="centerCrop">
-
-    <ImageButton
-        android:id="@+id/imageButtonVolumen"
-        android:layout_width="63dp"
-        android:layout_height="63dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
-        android:contentDescription="@string/botonvolumen"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/iconovolumenencendido" />
-
-    <ImageButton
-        android:id="@+id/imageButtonVolver"
-        android:layout_width="63dp"
-        android:layout_height="63dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="16dp"
-        android:background="@drawable/button_background"
-        android:contentDescription="@string/botonvolver"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/flecha"
-        tools:ignore="DuplicateSpeakableTextCheck" />
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="358dp"
-        android:layout_height="511dp"
-        android:layout_marginStart="46dp"
-        android:layout_marginEnd="47dp"
-        android:layout_marginBottom="160dp"
-        android:contentDescription="@string/fondo_blanco"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/fondo_blanco" />
-
-    <TextView
-        android:id="@+id/textViewContrasena"
-        android:layout_width="141dp"
-        android:layout_height="36dp"
-        android:text="@string/contrasenha"
-        android:textAlignment="textStart"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/editTextContrasena"
-        app:layout_constraintEnd_toEndOf="@+id/imageView"
-        app:layout_constraintHorizontal_bias="0.069"
-        app:layout_constraintStart_toStartOf="@+id/imageView"
-        app:layout_constraintTop_toBottomOf="@+id/editTextUsuario"
-        app:layout_constraintVertical_bias="0.593"
-        tools:ignore="TextSizeCheck" />
-
-    <EditText
-        android:id="@+id/editTextUsuario"
-        android:layout_width="326dp"
-        android:layout_height="50dp"
-        android:hint="@string/usuario"
-        android:inputType="textPersonName"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView"
-        app:layout_constraintEnd_toEndOf="@+id/imageView"
-        app:layout_constraintHorizontal_bias="0.468"
-        app:layout_constraintStart_toStartOf="@+id/imageView"
-        app:layout_constraintTop_toTopOf="@+id/imageView"
-        app:layout_constraintVertical_bias="0.206"
-        tools:ignore="Autofill,DuplicateSpeakableTextCheck,TextContrastCheck" />
-
-    <TextView
-        android:id="@+id/textViewUsuario"
-        android:layout_width="118dp"
-        android:layout_height="36dp"
-        android:text="@string/usuario"
-        android:textAlignment="textStart"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/editTextUsuario"
-        app:layout_constraintEnd_toEndOf="@+id/imageView"
-        app:layout_constraintHorizontal_bias="0.062"
-        app:layout_constraintStart_toStartOf="@+id/imageView"
-        app:layout_constraintTop_toTopOf="@+id/imageView"
-        app:layout_constraintVertical_bias="0.779"
-        tools:ignore="TextSizeCheck" />
-
-    <EditText
-        android:id="@+id/editTextContrasena"
-        android:layout_width="326dp"
-        android:layout_height="50dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="68dp"
-        android:layout_marginEnd="32dp"
-        android:hint="@string/contrasenha"
-        android:inputType="textPassword"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.428"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/editTextUsuario"
-        tools:ignore="Autofill,DuplicateSpeakableTextCheck" />
-
-    <Button
-        android:id="@+id/buttonRegister"
-        android:layout_width="159dp"
-        android:layout_height="61dp"
-        android:text="@string/registrarse"
-        android:backgroundTint="#FFB4A2"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/editTextContrasena"
-        app:layout_constraintVertical_bias="0.148" />
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="70dp"
-        android:layout_height="59dp"
-        android:layout_marginStart="144dp"
-        android:layout_marginBottom="60dp"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView"
-        app:layout_constraintStart_toStartOf="@+id/imageView"
-        app:srcCompat="@drawable/logo_google" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
