+++ app/src/main/res/layout/activity_results.xml
@@ -15,7 +15,8 @@
         android:layout_height="63dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:background="@drawable/button_background"
+        android:paddingBottom="8dp"
+        android:background="@drawable/button_square_background"
         android:contentDescription="@string/botonvolumen"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -25,8 +26,12 @@
         android:id="@+id/buttonMenu"
         android:layout_width="280dp"
         android:layout_height="252dp"
-        android:backgroundTint="#FFB4A2"
+        android:background="@drawable/button_background"
+        android:fontFamily="@font/bungee_regular"
         android:text="@string/menu"
+        android:textSize="18sp"
+        android:paddingBottom="9dp"
+        android:textColor="#6F5D47"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.495"
@@ -38,6 +43,7 @@
         android:id="@+id/textView"
         android:layout_width="263dp"
         android:layout_height="84dp"
+        android:fontFamily="@font/bungee_regular"
         android:text="Congrats!"
         android:textAlignment="center"
         android:textSize="60sp"
