+++ app/src/main/res/layout/activity_setting.xml
@@ -34,17 +34,4 @@
         app:srcCompat="@drawable/flecha"
         tools:ignore="DuplicateSpeakableTextCheck" />
 
-    <Button
-        android:id="@+id/buttonCerrarSesion"
-        android:layout_width="191dp"
-        android:layout_height="100dp"
-        android:layout_marginStart="110dp"
-        android:layout_marginEnd="110dp"
-        android:layout_marginBottom="548dp"
-        android:backgroundTint="#FFB4A2"
-        android:text="@string/cerrar_sesion"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
