+++ app/src/main/res/values/strings.xml
@@ -24,4 +24,5 @@
     <string name="oceania">Oceanía</string>
     <string name="click_me_to_win">Click me to win</string>
     <string name="menu">Menu</string>
+    <string name="logogoogle">logoGoogle</string>
 </resources>
